start supporting php.ini files
So, I want to make `-d` work from the command line. I tried to make a `IniSettings` section of the `hdf` but it turns out you can't have `.`s in any of the names or else it makes it a nested config section and `.get()` won't traverse it. Supporting a `IniFile` was much easier.

I set a default in `/etc/hhvm/php.ini`. Thoughts?

Reviewed By: @markw65

Differential Revision: D1009092