Some {W,}Iter{Init,Next}{K,} spec tweaks
The spec claimed these instructions stored to locals "as a
cell", which sounded to me like it should convert KindOfRef locals
into KindOfCell.  What it meant is that it stored using a normal
assignment (we don't really specify how these work, so I just changed
it to say "with the semantics of SetL").  Also hoist discussion of the
assignment semantics out of the switch (%1) part of the spec.

Reviewed By: @edwinsmith

Differential Revision: D1009178