Miscellaneous fixes: add newlines back, remove specific user path info, and help message update
Removing the final newlines from the expect files was premature and probably a mistake.

http://stackoverflow.com/questions/729692/why-should-files-end-with-a-newline

So, leave it to clients/upstream to deal with newlines at the end of the file.

Also, do not print user path info out in expect files (or other files). e.g., no joelm

Also, list frameworks in help message when specifying invalid ones

Reviewed By: @ptarjan

Differential Revision: D1045907